+++ plugins/Moodle/School_Setup/Calendar.php
@@ -138,7 +138,7 @@ function core_calendar_create_calendar_events_response( $response )
 		$calendar_event_id = $_REQUEST['event_id'];
 	}
 
-	DBQuery( "INSERT INTO MOODLEXROSARIO (\"column\",rosario_id,moodle_id)
+	DBQuery( "INSERT INTO MOODLEXROSARIO (" . DBEscapeIdentifier( 'column' ) . ",rosario_id,moodle_id)
 		VALUES('calendar_event_id', '" . $calendar_event_id . "', " . $response['events'][0]['id'] . ")" );
 
 	return null;
@@ -193,7 +193,7 @@ function core_calendar_delete_calendar_events_response( $response )
 
 	//delete the reference the moodlexrosario cross-reference table:
 	DBQuery( "DELETE FROM MOODLEXROSARIO
-		WHERE \"column\"='calendar_event_id'
+		WHERE " . DBEscapeIdentifier( 'column' ) . "='calendar_event_id'
 		AND rosario_id='" . $rosario_id . "'" );
 
 	return null;